+++ crates/rattler_conda_types/Cargo.toml
@@ -21,7 +21,6 @@ nom = "7.1.3"
 regex = "1.9.4"
 serde = { version = "1.0.188", features = ["derive"] }
 serde_json = "1.0.105"
-serde-json-python-formatter = "0.1.0"
 serde_yaml = "0.9.25"
 serde_with = { version = "3.3.0", features = ["indexmap_2"] }
 serde_repr = "0.1"
