+++ crates/rattler_lock/Cargo.toml
@@ -1,0 +1,30 @@
+[package]
+name = "rattler_lock"
+version.workspace = true
+edition.workspace = true
+authors = ["Bas Zalmstra <zalmstra.bas@gmail.com>"]
+description = "Rust data types for conda lock"
+categories.workspace = true
+homepage.workspace = true
+repository.workspace = true
+license.workspace = true
+readme.workspace = true
+
+[dependencies]
+chrono = "0.4.27"
+fxhash = "0.2.1"
+indexmap = { version = "2.0.0", features = ["serde"] }
+rattler_conda_types = { version = "0.10.0", path = "../rattler_conda_types" }
+rattler_digest = { version = "0.10.0", path = "../rattler_digest" }
+pep508_rs = { version = "0.2.3", features = ["serde"] }
+pep440_rs = { version = "0.3.12", features = ["serde"] }
+serde = { version = "1.0.188", features = ["derive"] }
+serde-json-python-formatter = "0.1.0"
+serde_json = "1.0.105"
+serde_yaml = "0.9.25"
+serde_with = { version = "3.3.0", features = ["indexmap_2"] }
+thiserror = "1.0.47"
+url = { version = "2.4.1", features = ["serde"] }
+
+[dev-dependencies]
+insta = { version = "1.31.0", features = ["yaml"] }
