+++ crates/rattler_lock/src/utils/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod serde;
